+++ src/server/products/getTag.ts
@@ -1,4 +1,6 @@
-export async function getTag(field, value) {
+import { Tag } from '@typings/products/tag';
+
+export async function getTag<K extends 'id' | 'slug'>(field: K, value: Tag[K]): Promise<Tag | undefined> {
   try {
     const res = await fetch('/data/product-tags.json');
 
@@ -6,10 +8,8 @@ export async function getTag(field, value) {
       throw new Error(`Failed to fetch product-tags.json: ${res.status}`);
     }
 
-    const tags = await res.json();
+    const tags: Tag[] = await res.json();
     const tag = tags.find((tag) => tag[field] === value);
-
-    if (!tag) return;
 
     return tag;
   } catch (error) {
@@ -18,10 +18,10 @@ export async function getTag(field, value) {
   }
 }
 
-export function getTagById(id) {
+export function getTagById(id: number) {
   return getTag('id', id);
 }
 
-export function getTagBySlug(slug) {
+export function getTagBySlug(slug: string) {
   return getTag('slug', slug);
 }